Properly parse JSON in caching logic (#71)
Hishel works with httpcore.Response, not httpx.Response, which doesn't
do e.g. gzip decoding. In order to correctly parse the JSON we have to
construct an httpx.Response (which Hishel will do later but not at the
point where we need it).

Without this, responses containing a JSON list of errors are incorrectly
cached.